requirements.txt
PyPDF2==1.26.0
reportlab==3.6.1
Pillow==8.4.0
PyMuPDF==1.18.16